+++ src/runtime/_tokio.rs
@@ -308,7 +308,7 @@ pub trait ClientLike: Clone + Send + Sync + Sized {
         }
       }
 
-      utils::check_and_set_client_state(&inner.state, ClientState::Disconnecting, ClientState::Disconnected);
+      inner.cas_client_state(ClientState::Disconnecting, ClientState::Disconnected);
       result
     })
   }